module-04/exercises/05-jwt-in-flask--complete/requirements.txt
click==7.1.1
ecdsa==0.15
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pyasn1==0.4.8
python-jose==3.1.0
rsa==4.0
six==1.14.0
Werkzeug==1.0.1
